rename encoding.rs and refactor some
