+++ contracts/multisig-prover/src/execute.rs
@@ -12,7 +12,7 @@ use service_registry::state::Worker;
 
 use crate::{
     contract::START_MULTISIG_REPLY_ID,
-    encoding::encoding::CommandBatchBuilder,
+    encoding::CommandBatchBuilder,
     error::ContractError,
     events::Event,
     state::{
@@ -33,7 +33,8 @@ pub fn construct_proof(deps: DepsMut, message_ids: Vec<String>) -> Result<Respon
     let command_batch = match COMMANDS_BATCH.may_load(deps.storage, &batch_id)? {
         Some(batch) => batch,
         None => {
-            let mut builder = CommandBatchBuilder::new(config.destination_chain_id, config.encoding_scheme);
+            let mut builder =
+                CommandBatchBuilder::new(config.destination_chain_id, config.encoding);
             for msg in messages {
                 builder.add_message(msg)?;
             }
@@ -139,7 +140,7 @@ pub fn update_worker_set(deps: DepsMut, env: Env) -> Result<Response, ContractEr
     }
 
     NEXT_WORKER_SET.save(deps.storage, &new_worker_set)?;
-    let mut builder = CommandBatchBuilder::new(config.destination_chain_id, config.encoding_scheme);
+    let mut builder = CommandBatchBuilder::new(config.destination_chain_id, config.encoding);
     builder.add_new_worker_set(new_worker_set)?;
 
     let batch = builder.build()?;
