+++ contracts/multisig-prover/src/msg.rs
@@ -4,7 +4,7 @@ use axelar_wasm_std::Threshold;
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{HexBinary, Uint256, Uint64};
 
-use crate::encoding::{Data, encoding::EncodingScheme};
+use crate::encoding::{Data, EncodingScheme};
 
 #[cw_serde]
 pub struct InstantiateMsg {
@@ -17,9 +17,8 @@ pub struct InstantiateMsg {
     pub service_name: String,
     pub chain_name: String,
     pub worker_set_diff_threshold: u32,
-    pub encoding_scheme: EncodingScheme
+    pub encoding: EncodingScheme,
 }
-
 
 #[cw_serde]
 pub enum ExecuteMsg {
