+++ contracts/multisig-prover/src/types.rs
@@ -6,7 +6,10 @@ use cw_storage_plus::{Key, KeyDeserialize, PrimaryKey};
 use multisig::key::Signature;
 use sha3::{Digest, Keccak256};
 
-use crate::{encoding::{Data, encoding::EncodingScheme}, state::WorkerSet};
+use crate::{
+    encoding::{Data, EncodingScheme},
+    state::WorkerSet,
+};
 
 #[cw_serde]
 pub enum CommandType {
@@ -81,7 +84,7 @@ pub struct CommandBatch {
     pub id: BatchID,
     pub message_ids: Vec<String>,
     pub data: Data,
-    pub encoding_scheme: EncodingScheme
+    pub encoding: EncodingScheme,
 }
 
 #[cw_serde]
